requirements.txt
requests>=2.28.0
python-dotenv>=0.20.0
colorama>=0.4.5